% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distR.R
\name{CalcLaplacianMatrix}
\alias{CalcLaplacianMatrix}
\title{Calculate the Laplacian matrix of a linear network}
\usage{
CalcLaplacianMatrix(L)
}
\arguments{
\item{L}{A linear network of class linnet from the spatstat package.}
}
\value{
A matrix.
}
\description{
Calculates the Laplacian matrix of a linear network from, and adds a +1
to the first entry in the matrix to make it invertible.
}
\examples{
CalcLaplacianMatrix(spatstat.data::simplenet)

L = as.linnet(spatstat.data::spiders)
CalcLaplacianMatrix(L)

L = as.linnet(spatstat.data::dendrite,sparse=FALSE)
CalcLaplacianMatrix(L)
}
\seealso{
\link{distR} and \link{pairdistR} for calculation of the resistance
metric using the Laplacian matrix
}
