% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfunc.R
\name{covfunctypes}
\alias{covfunctypes}
\title{Create a covariance function}
\usage{
covfunctypes(type, par = NULL)
}
\arguments{
\item{type}{The types implemented are "expcov", "gamma", "invgamma"
or "gig"; see the details below.}

\item{par}{if par=NULL (the default behavior), the parameters
can be specified later; otherwise this should be a vector containing
the correct number of parameters; see the details below.}
}
\value{
A function of distance r, or if par=NULL, a function of
(par,r) where par is the parameter vector.
}
\description{
Creates an isotropic covariance (or, more precisely, correlation)
function for use on linear networks.
}
\details{
The functions returns a covariance (or correlation) function.
If par=NULL, this is a function of (par,r) where par is a
parameter vector, while if par is a vector it returns a function
of r. A supplied parameter vector should have the correct number
of entries for the chosen covariance function. The currently
implemented covariance functions are:
\itemize{
\item Exponential covariance function: c0(r) = exp(-sr),
with a scale parameter s>0. Set type="expcov" and par=c(s).
\item Covariance function with gamma Bernstein CDF:
c0(r) = (1+r/phi)^(-tau), with shape parameter tau and rate
parameter phi. Set type="gamma" and par = c(tau,phi).
\item Covariance function with inverse gamma Bernstein CDF:
c0(r) = 2(r phi)^(tau/2)K_tau(2 sqrt(r phi)/(Gamma(tau)),
with shape parameter tau and scale
parameter phi. Set type="invgamma" and par = c(tau,phi).
\item Covariance function with generalized inverse Gaussian Bernstein
CDF: c0(r) = (1+2r/psi)^(-lambda/2)(K_lambda(sqrt((2r+psi)chi))/
(K_lambda(sqrt(psi chi))
with parameters (chi,psi,lambda); set type="gig" and
par = c(chi,psi,lambda).
}
}
\examples{
# Exponential covariance function with scale parameter 1
c0 = covfunctypes("expcov",1)
curve(c0(x),from=0,to=5)

# Exponential covariance function with unspecified parameter
c0 = covfunctypes("expcov")
curve(c0(1,x),from=0,to=5)  # scale=1 specified here

# Covariance function with gamma Bernstein CDF
c0 = covfunctypes("gamma",par=c(1,0.5))  # tau=1 and phi=0.5
curve(c0(x),from=0,to=5)

# Covariance function with inverse gamma Bernstein CDF
c0 = covfunctypes("invgamma",par=c(1,2))  # tau=1 and phi=2
curve(c0(x),from=0,to=5)

# Covariance function with generalized inverse Gaussian Bernstein CDF
c0 = covfunctypes("gig",par=c(1,1,1))  # (chi,psi,lambda) = (1,1,1)
curve(c0(x),from=0,to=5)
}
\seealso{
\link{paircorfunc} for pair correlation functions, \link{simalgotypes}
for simulation of Bernstein distributions corresponding to covariance
functions, \link{simGausLNDisc} for simulation of GRFs on linear networks
with a given covariance function
}
