% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxln-package.R
\docType{package}
\name{coxln-package}
\alias{coxln}
\alias{coxln-package}
\title{coxln: Cox processes on Linear Networks}
\description{
Simulation algorithms for Cox processes driven by transformed Gaussian processes on linear networks, and other tools needed for analysing data with these processes.
}
\author{
\strong{Maintainer}: Rasmussen Jakob G. \email{jgr@math.aau.dk}

}
\keyword{internal}
