% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearFGJ.R
\name{findendpoints}
\alias{findendpoints}
\title{Find end points in network}
\usage{
findendpoints(L)
}
\arguments{
\item{L}{A linear network.}
}
\value{
A point pattern on a linear network with class lpp
from spatstat.
}
\description{
Finds all the end points in a network, i.e. all the vertices
of degree one in the graph underlying the network.
}
\details{
Given a linear network, this function returns a point pattern
on the same network giving all the end points in the network.
}
\examples{
# Find the end points from simplenet from spatstat
ep = findendpoints(simplenet)
plot(ep)
}
\seealso{
\link{minussampling} for a minus sampling method using
the end points
}
