% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distR.R
\name{getdiag}
\alias{getdiag}
\title{Extract diagonal from a matrix}
\usage{
getdiag(x)
}
\arguments{
\item{x}{A square matrix.}
}
\value{
A vector with diagonal entries from x.
}
\description{
Returns the diagonal of a matrix, but unlike diag() it considers a
single number to be matrix.
}
