% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGaus.R
\name{is.tree}
\alias{is.tree}
\title{Check whether L is a connected tree}
\usage{
is.tree(L)
}
\arguments{
\item{L}{A linear network}
}
\value{
Returns TRUE is L is both connected and a tree; returns FALSE otherwise.
}
\description{
Checks whether L is a connected tree.
}
\examples{
# Check whether simplenet from spatstat is a connected tree
is.tree(spatstat.data::simplenet)

# Check whether the network used in the dendrite data is a connected tree
is.tree(as.linnet(spatstat.data::dendrite))
}
