% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearFGJ.R
\name{linearJ}
\alias{linearJ}
\title{Estimate J function on linear network}
\usage{
linearJ(X, r, ppul, metric = "G")
}
\arguments{
\item{X}{A point pattern on a linear network.}

\item{r}{A vector of distances on which the estimate of J(r) is calculated.}

\item{ppul}{The number of points per unit length of the linear network used
for approximating the F function used in the J-function.
This is rounded down to an integer for each line segment.
Note that the vertices in the graph underlying the linear network are
automatically added, and are not counted in ppul.}

\item{metric}{The metric used to measure distances. Should be either "G"
for geodesic distance, or "R" for resistance distance.}
}
\value{
The J function with class fv from spatstat.
}
\description{
Estimates the J function for a point pattern observed on a linear network.
}
\details{
This function takes a point pattern on a linear network and estimates
the J function for a set of r-values by discretising the network. Both the
geodesic and the resistance metric can be used for distances
}
\examples{
# Estimate J function for spiders data
X = spatstat.data::spiders
r = seq(0,250,length.out=50)
ppul = 0.02
J = linearJ(X,r,ppul,metric="G")
plot(J)

# Estimate J function for dendrite data
X = spatstat.data::dendrite
r = seq(0,5,length.out=50)
ppul = 0.5
J = linearJ(X,r,ppul,metric="G")
plot(J)
}
\seealso{
\link{linearF}, \link{linearG} and \link{linearpcfR} for other summary statistics,
\link{minussampling} for minus sampling
}
