% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearpcfR.R
\name{linearpcfR}
\alias{linearpcfR}
\title{Estimate pcf using resistance metric}
\usage{
linearpcfR(X, r, silent = TRUE)
}
\arguments{
\item{X}{A point pattern on a linear network.}

\item{r}{A vector of distances on which the estimate of the pair correlation
function is calculated.}

\item{silent}{If FALSE it outputs how far the algorithm has progressed during calculations;
default is TRUe.}
}
\value{
The J function with class fv from spatstat.
}
\description{
Estimates the pair correlation function from a point pattern
on a linear network using the resistance metric.
}
\details{
This function takes a point pattern on a linear network and estimates
the pair correlation function for a set of r-values by discretising the network.
The resistance metric is used for distances.
}
\examples{
# Estimate pcf for spiders data
X = spatstat.data::spiders
r = seq(0,250,length.out=50)
pcf = linearpcfR(X,r)
plot(pcf)
}
\seealso{
\link{linearF}, \link{linearG} and \link{linearJ} for other summary statistics
}
