% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGaus.R
\name{makeorderL}
\alias{makeorderL}
\title{Make order on line segments}
\usage{
makeorderL(L, orderV = makeorderV(L))
}
\arguments{
\item{L}{A linear network}

\item{orderV}{A vector of orders on the vertices; created by the functino makeorderV by
default.}
}
\value{
Returns a vector of integers representing the order of each line segment.
}
\description{
Makes an order of the line segments in a linear network by giving the same order
to a line segments as the lowest order of the vertices it connects to.
}
\examples{
# Make order on the vertices from simplenet
makeorderL(simplenet)

# Make order on the vertices in the network used in the dendrite data
makeorderL(as.linnet(dendrite))
}
\seealso{
\link{makeorderV} for orders on vertices, \link{simGausExpLNRoot} or
\link{simGausLNRoot} for simulation of GRFs using orders, \link{simCPExpLNRoot} or
\link{simCPLNRoot} for simulation of Cox processes using orders
}
