% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGaus.R
\name{makeorderV}
\alias{makeorderV}
\title{Make order on vertices}
\usage{
makeorderV(L)
}
\arguments{
\item{L}{A linear network}
}
\value{
Returns a vector of integers representing the order of each vertex.
}
\description{
Makes an order of the vertices in a linear network (defined by the smallest
number of line segments from the vertex internally numbered as 1).
}
\examples{
# Make order on the vertices from simplenet
makeorderV(spatstat.data::simplenet)

# Make order on the vertices in the network used in the dendrite data
makeorderV(as.linnet(spatstat.data::dendrite))
}
\seealso{
\link{makeorderL} for orders on line segments, \link{simGausExpLNRoot} or
\link{simGausLNRoot} for simulation of GRFs using orders, \link{simCPExpLNRoot} or
\link{simCPLNRoot} for simulation of Cox processes using orders
}
