% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearFGJ.R
\name{minussampling}
\alias{minussampling}
\title{Removes points within distance r of end points}
\usage{
minussampling(X, r, metric = "G", saveind = FALSE)
}
\arguments{
\item{X}{A point pattern on a linear network.}

\item{r}{The distance to the end points, within which points should be
removed.}

\item{metric}{The metric used to measure distances. Should be either "G"
for geodesic distance, or "R" for resistance distance.}

\item{saveind}{For internal use.}
}
\value{
A point pattern on a linear network with class lpp from spatstat.
}
\description{
Removes all points within distance r from the end points of
a point pattern defined on a linear network.
}
\details{
This function takes a point pattern X on a linear network, and
removes all points within distance r from the end points of the
network, i.e. all the vertices of degree one in the graph
underlying the network. The distances can be measured either by
geodesic distance or resistance distance.
}
\examples{
# Removes points from the dendrite dataset using geodesic metric
X = spatstat.data::dendrite
r = 20
Xm = minussampling(X,r,metric="G")
plot(X)
plot(Xm, add=TRUE, cols=3)

# Removes points from the spiders dataset using resistance metric
X = spatstat.data::spiders
r = 200
Xm = minussampling(X,r,metric="R")
plot(X)
plot(Xm, add=TRUE, cols=3)
}
