% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfunc.R
\name{paircorfunc}
\alias{paircorfunc}
\title{Create a pair correlation function}
\usage{
paircorfunc(covtype, transform)
}
\arguments{
\item{covtype}{The type of covariance function used;
currently available: "expcov", "gamma", "invgamma" and "gig";
see details under the function covfunctypes.}

\item{transform}{The type of point process used; "lgcp" for
log Gaussian Cox process, "icp" for interrupted Cox process, and
"pcpp" for permanental Cox process.}
}
\value{
A function with inputs par and r.
}
\description{
This creates a theoretical pair correlation function for a
LGCP, ICP or PCPP model with a chosen covariance function
for the underlying Gaussian process.
}
\details{
This function creates a pair correlation function with inputs
par and r, where par is a vector of parameters used both in the
chosen point process, and the chosen covariance function, and r
is the distance at with the pair correlation function is evaluated.
The parameters in par depend on the point process and the covariance
function, where the first parameters are used by the choice of
point process as follow:
\itemize{
\item "lgcp": par = (sigma, ...)
\item "icp": par = (sigma, h, ...)
\item "pcpp": par = (h, ...)
}

The remaining parameters in the par vector is used by the
covariance function (see the covfunctypes function). Note that
the number of parameters in the par vector should fit the total
number of parameters used by the point process and the covariance
function.
}
\examples{
# pcf for LGCP with exponential covariance function
pcf = paircorfunc("expcov","lgcp")
parameters = c(sigma=1,s=1)
curve(pcf(parameters,x),from=0,to=5,ylab="pcf",xlab="r")

# pcf for ICP with gamma Bernstein CDF
pcf = paircorfunc("gamma","icp")
parameters = c(sigma=1,h=1,tau=1,phi=1)
curve(pcf(parameters,x),from=0,to=5,ylab="pcf",xlab="r")

# pcf for PCPP with inverse gamma Bernstein CDF
pcf = paircorfunc("invgamma","pcpp")
parameters = c(h=1,tau=1,phi=1)
curve(pcf(parameters,x),from=0,to=5,ylab="pcf",xlab="r")
}
\seealso{
\link{covfunctypes} for covariance functions, \link{simGausLNDisc}
for simulation of GRFs on linear networks
with a given covariance function, \link{parnamespcf} for names of parameters
in the pair correlation function
}
