% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distR.R
\name{pairdistR}
\alias{pairdistR}
\title{Pairwise distances of point pattern on linear network using the resistance metric}
\usage{
pairdistR(
  X,
  X2 = NULL,
  derivative = FALSE,
  L = as.linnet(X),
  ILM = solve(CalcLaplacianMatrix(L))
)
}
\arguments{
\item{X}{If a single point pattern is provided all pairwise distances (using the
resistance distance) between points in X are found.}

\item{X2}{If a second point pattern is provided, all pairwise distances will be found
between points in X and X2 are found.}

\item{derivative}{If TRUE then the derivative of the resistance distance d(u,v) wrt. v
is found instead of the resistance distance itself. Default is FALSE.}

\item{L}{Internal use.}

\item{ILM}{Internal use.}
}
\value{
A matrix with all the distances or derivatives of the distances.
}
\description{
Calculates all pairwise distances using the resistance distance between all points in
a point pattern or two patterns on a linear network. Can also calculate the derivative
(i.e. Jacobian) instead.
}
\examples{
# All pairwise distances in the spiders dataset
X = spatstat.data::spiders
pairdistR(X)

# Sparse representations in the linear network does not work
X = spatstat.data::dendrite
X = as.lpp(X,sparse=FALSE)
pairdistR(X)

# All pairwise distances between two Poisson processes on simplenet
L = spatstat.data::simplenet
X = rpoislpp(5,L)
Y = rpoislpp(5,L)
pairdistR(X,Y)

# The derivative of the resistance distance
X = spatstat.data::spiders
pairdistR(X,derivative=TRUE)
}
