% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfunc.R
\name{parnamespcf}
\alias{parnamespcf}
\title{Output parameter vector of a pair correlation function}
\usage{
parnamespcf(covtype, transform = "none")
}
\arguments{
\item{covtype}{The type of covariance function used;
currently available: "expcov", "powexpcov", "materncov", "gencaucov", "dagumcov",
"gammabd", "invgammabd" and "gigbd";
see details under the function covfunctypes.}

\item{transform}{The type of point process used; "lgcp" for
log Gaussian Cox process, "icp" for interrupted Cox process, and
"pcpp" for permanental Cox process. If transform = "none" only the
parameters in the covariance function is returned.}
}
\value{
A vector of strings with the names of the parameters.
}
\description{
Since the pair correlation function both gets parameters from the covariance function
and the class of point process used, the number and order of parameters in its
parameter vector can be confusing. This function outputs the names of parameters in the
parameter vector of the pair correlation function.
}
\examples{
# parameters in pcf for LGCP with exponential covariance function
parnamespcf("expcov","lgcp")

# parameters in pcf for ICP with gamma covariance function
parnamespcf("gammabd","icp")

# parameters in covariance function with generalized inverse Gaussian as Bernstein distribution
parnamespcf("gigbd")
}
\seealso{
\link{paircorfunc} for pair correlation functions, \link{covfunctypes}
for covariance functions.
}
