% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearpcfR.R
\name{relativeboundaryR}
\alias{relativeboundaryR}
\title{Find the relative boundary using resistance distance}
\usage{
relativeboundaryR(
  point,
  r,
  L = as.linnet(point),
  ILM = solve(CalcLaplacianMatrix(L)),
  outpp = TRUE
)
}
\arguments{
\item{point}{A point pattern on a linear network consisting of a single point.}

\item{r}{The distance from the point to the relative boundary.}

\item{L}{For internal use.}

\item{ILM}{For internal use.}

\item{outpp}{If TRUE the a point pattern is returned; otherwise a dataframe is returned.}
}
\value{
Either a point pattern of type lpp from spatstat or a dataframe.
}
\description{
Find the relative boundary from a given point on a linear network
at a given resistance distance.
}
\details{
This function finds the relative boundary from a point in a linear network,
and returns this as a point pattern on the same network. It uses the resistance
metric for calculating distances.
}
\examples{
# Relative boundary from point no. 30 in the spiders data
X = spatstat.data::spiders
point = X[30]
r = 180
rb = relativeboundaryR(point,r)
plot(point)
plot(rb,add=TRUE,cols=2)
# Relative boundary from point no. 10 in the chicago data
X = spatstat.data::chicago
point = X[10]
r = 100
rb = relativeboundaryR(point,r)
plot(point,use.marks=FALSE)
plot(rb,add=TRUE,cols=2)
}
\seealso{
\link{linearpcfR} for estimation of the pair correlation function
using the relative boundary for edge correction
}
