% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covfunc.R
\name{simalgotypes}
\alias{simalgotypes}
\title{Simulate a Bernstein density}
\usage{
simalgotypes(param, type, nsim)
}
\arguments{
\item{param}{Parameter vector used in the Bernstein distribution.}

\item{type}{The type of Bernstein distribution. Can be "gammabd",
"invgammabd", and "gigbd". See details.}

\item{nsim}{The number of simulations.}
}
\value{
A vector of length nsim with the simulated values.
}
\description{
Makes a simulation from the Bernstein density corresponding to
a covariance function.
}
\details{
This function makes a number of simulations from the Bernstein density
corresponding to a covariance function (see the function covfunc for details
on the corresponding covariance functions). This vector can be used in the
function simGausLNRoot for making approximate simulations of a Gaussian random
field with the corresponding covariance function. The currently implemented
types are:
\itemize{
\item Gamma Bernstein density: gamma distribution with shape parameter tau and rate
parameter phi. Set type="gammabd" and param = c(tau,phi).
\item Inverse Gamma Bernstein density: Inverse gamma distribution with shape
parameter tau and scale parameter phi. Set type="invgammabd" and param = c(tau,phi).
\item Generalized inverse Gaussian Bernstein density: Generalized inverse Gaussian
distribution with with parameters (chi,psi,lambda). Set type="gigbd" and
param = c(chi,psi,lambda).
}
}
\examples{
# simulation of gamma Bernstein distribution
simalgotypes(param=c(1,1),type="gammabd",nsim=10)

# simulation of inverse gamma Bernstein distribution
simalgotypes(param=c(1,1),type="invgammabd",nsim=10)

# simulation of generalized inverse Gaussian Bernstein distribution
simalgotypes(param=c(1,1,1),type="gigbd",nsim=10)

}
\seealso{
\link{covfunctypes} for covariance functions corresponding to
Bernstein distributions, \link{simGausExpLNRoot} or
\link{simGausLNRoot} for simulation of GRFs on tree-shaped linear networks
with a given Bernstein distribution
}
