% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearpcfR.R
\name{wR}
\alias{wR}
\title{Calculate weights used in estimation of pcf}
\usage{
wR(u, r, L = as.linnet(u), ILM = solve(CalcLaplacianMatrix(L)))
}
\arguments{
\item{u}{A point pattern on a linear network consisting of a single point.}

\item{r}{A distance.}

\item{L}{For internal use.}

\item{ILM}{For internal use.}
}
\value{
A real number representing the weight.
}
\description{
Calculates the weights used in estimation of the pair correlation
function for a point pattern observed on a linear network using the
resistance metric.
}
\seealso{
\link{linearpcfR} for estimation of the pair correlation function
using the resistance metric
}
