% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGaus.R
\name{makeorderL}
\alias{makeorderL}
\title{Make order line segments}
\usage{
makeorderL(L, orderV = makeorderV(L))
}
\arguments{
\item{L}{A linear network}

\item{orderV}{A vector of orders on the vertices; created by the functino makeorderV by
default.}
}
\value{
Returns a vector of integers representing the order of each line segment.
}
\description{
Makes an order of the line segments in a linear network by giving the same order
to a line segments as the lowest order of the vertices it connects to.
}
\examples{
# Make order on the vertices from simplenet
makeorderL(spatstat.data::simplenet)

# Make order on the vertices in the network used in the dendrite data
makeorderL(as.linnet(spatstat.data::dendrite))
}
