% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGaus.R
\name{simY}
\alias{simY}
\title{Make simulations for simGausExpLNRoot}
\usage{
simY(pos, beta, mu, sigma, orderL, orderV)
}
\description{
This is an internal function for making the simulations
used in the function simGausExpLNRoot.
}
